Refactor location_buffer into new file
