+++ src/canvas.h
@@ -7,6 +7,7 @@
 
 struct pipe;
 struct pipe_cell_list;
+struct location_buffer;
 
 struct palette {
     uint32_t *colors;
@@ -34,20 +35,6 @@ cpipes_errno canvas_resize(
 void canvas_free(struct canvas *canvas);
 void canvas_erase_tail(struct canvas *c,
         unsigned int tail, struct pipe *p);
-
-// Circular buffer containing the locations of a pipe
-struct location_buffer {
-    unsigned int *locations;
-    size_t max, size, head;
-};
-
-cpipes_errno location_buffer_init(struct location_buffer *buffer, size_t max);
-void location_buffer_free(struct location_buffer *buffer);
-
-unsigned int *location_buffer_head(struct location_buffer *buffer);
-unsigned int *location_buffer_tail(struct location_buffer *buffer);
-
-void location_buffer_push(struct location_buffer *buffer, unsigned int loc);
 
 #endif /* CANVAS_H_ */
